test(schedule-engine): split scheduleEngine.test.ts so its two heavy cases shard apart
